Updated the tracker to treat DEM uncertainty properly.  In particular, the particles are no longer explicitly tied to the glacier surface.  Rather, the z dimension is now subject to a motion model, just as the map-plane dimensions are, with the DEM entering through a new component of the likelihood model which provides higher weights to particles that are close to the observed glacier surface.
More should be done with this.  In particular, we need to adjust tracker to take a set of DEMs with weights, which will allow implicitly for easy DEM time-interpolation.  Also, the new DEM argument Z_sigma should be modified to be a spatial interpolant just like the elevation itself.  Thus, particles that are in regions of the DEM that are not believable should not be constrained to lie on the surface.